+++ inlang/source-code/cli/src/commands/project/validate.ts
@@ -14,12 +14,12 @@ export async function validateCommandAction() {
 		// Get the config
 		const { error } = await getInlangProject()
 		if (error) {
-			log.error(`❌ ${error} (${error.message})`)
+			log.error(`${error} (${error.message})`)
 			return
 		}
 
-		log.success("🎉 Inlang config is valid!")
+		log.success("Inlang config is valid!")
 	} catch (error) {
-		log.error(`❌ ${error}`)
+		log.error(error)
 	}
 }