+++ inlang/source-code/website/src/pages/document/index.page.server.tsx
@@ -1,0 +1,17 @@
+const page = {
+	slug: "document",
+	content: {
+		title: "Global Document",
+		description: "Documents made to collaborate with version control and multi-language support.",
+	},
+}
+
+export async function onBeforeRender() {
+	return {
+		pageContext: {
+			pageProps: {
+				...page,
+			},
+		},
+	}
+}