+++ lix/source-code/client/src/index.ts
@@ -1,3 +1,4 @@
 export { type Repository } from "./api.js"
 export { openRepository } from "./openRepository.js"
 export { createNodeishMemoryFs } from "@lix-js/fs"
+export { default as browserAuth } from "./browser-auth.js"