+++ inlang/source-code/plugins/i18next/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @inlang/plugin-i18next
 
+## 4.8.0
+
+### Minor Changes
+
+- 0055f20b1: update README
+
+## 4.7.0
+
+### Minor Changes
+
+- 7bcb365ed: update `config init` deprecation
+
+### Patch Changes
+
+- 4668f637a: Added test for empty object in nested translation file.
+- Updated dependencies [2976a4b15]
+  - @inlang/sdk@0.10.0
+
 ## 4.6.0
 
 ### Minor Changes